+++ .idea/inspectionProfiles/Project_Default.xml
@@ -3,7 +3,7 @@
     <option name="myName" value="Project Default" />
     <inspection_tool class="DuplicatedCode" enabled="true" level="WEAK WARNING" enabled_by_default="true">
       <Languages>
-        <language minSize="77" name="Java" />
+        <language minSize="76" name="Java" />
       </Languages>
     </inspection_tool>
   </profile>
