+++ src/main/java/vn/edu/hcmuaf/fit/model/Product.java
@@ -6,9 +6,7 @@ public class Product {
     private double priceDiscount;
     private double price;
     private String image;
-    private String size;
     private int quantity;
-    private double weight;
 
     public Product() {
     }
@@ -20,33 +18,6 @@ public class Product {
         this.price = price;
         this.image = image;
         this.quantity = quantity;
-    }
-
-    public Product(int idBook, String name, double priceDiscount, double price, String image, String size, int quantity, double weight) {
-        this.idBook = idBook;
-        this.name = name;
-        this.priceDiscount = priceDiscount;
-        this.price = price;
-        this.image = image;
-        this.size = size;
-        this.quantity = quantity;
-        this.weight = weight;
-    }
-
-    public double getWeight() {
-        return weight;
-    }
-
-    public void setWeight(double weight) {
-        this.weight = weight;
-    }
-
-    public String getSize() {
-        return size;
-    }
-
-    public void setSize(String size) {
-        this.size = size;
     }
 
     public int getIdBook() {
