+++ src/main/java/vn/edu/hcmuaf/fit/services/IOrderService.java
@@ -1,11 +1,11 @@
 package vn.edu.hcmuaf.fit.services;
 
-import vn.edu.hcmuaf.fit.model.CartModel;
+import vn.edu.hcmuaf.fit.model.Cart;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 public interface IOrderService {
     boolean checkIdExistsInCart(String listId, HttpServletRequest request, HttpServletResponse response);
-    CartModel cartOrder(String list, HttpServletRequest request);
+    Cart cartOrder(String list, HttpServletRequest request);
 }
