+++ src/main/java/vn/edu/hcmuaf/fit/utils/FeeGHNUtils.java
@@ -20,14 +20,13 @@ import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.message.BasicNameValuePair;
 import org.apache.http.util.EntityUtils;
-import org.json.JSONArray;
 import org.json.JSONObject;
 import java.net.URL;
 import vn.edu.hcmuaf.fit.constant.GHNContent;
 
 public class FeeGHNUtils {
 
-	public static ArrayList<String> registerShip(int height, int lenght, int with,int weght,int  from_district_id, int from_ward_id,String tenHuyen, String tenXa) {
+	public static ArrayList<String> feeAPIAndDate(String height, String lenght, String with,String weght,int  from_district_id, int from_ward_id,String tenHuyen, String tenXa) {
 		String apiUrl = "http://140.238.54.136/api/registerTransport";
 		String authToken =token();
 		ArrayList<String> result = new ArrayList<>();
@@ -36,10 +35,10 @@ public class FeeGHNUtils {
 			int ward_id_to = DistrictGHNUtils.getDistrictIdOfWard(tenXa,districtId_to);
 			// Create the parameters for the request
 			Map<String, String> params = new HashMap<>();
-			params.put("height", height+"");
-			params.put("length", lenght+"");
-			params.put("width", with+"");
-			params.put("weight", weght+"");
+			params.put("height", height);
+			params.put("length", lenght);
+			params.put("width", with);
+			params.put("weight", weght);
 			params.put("from_district_id", from_district_id+"");
 			params.put("from_ward_id", from_ward_id+"");
 			params.put("to_district_id", districtId_to+"");
@@ -105,128 +104,6 @@ public class FeeGHNUtils {
 		return null;
 
 	}
-	public static double getFeeShip(int height, int lenght, int with,int weght,int  from_district_id, int from_ward_id,String tenHuyen, String tenXa) {
-		String apiUrl = "http://140.238.54.136/api/calculateFee";
-		String authToken =token();
-		double result = 0.0;
-		try {
-			int districtId_to = DistrictGHNUtils.getDistrictId(tenHuyen);
-			int ward_id_to = DistrictGHNUtils.getDistrictIdOfWard(tenXa,districtId_to);
-			// Create the parameters for the request
-			Map<String, String> params = new HashMap<>();
-			params.put("height", height+"");
-			params.put("length", lenght+"");
-			params.put("width", with+"");
-			params.put("weight", weght+"");
-			params.put("from_district_id", from_district_id+"");
-			params.put("from_ward_id", from_ward_id+"");
-			params.put("to_district_id", districtId_to+"");
-			params.put("to_ward_id", ward_id_to+"");
-			StringBuilder postData = new StringBuilder();
-			for (Map.Entry<String, String> param : params.entrySet()) {
-				if (postData.length() != 0) postData.append('&');
-				postData.append(URLEncoder.encode(param.getKey(), "UTF-8"));
-				postData.append('=');
-				postData.append(URLEncoder.encode(param.getValue(), "UTF-8"));
-			}
-			URL url = new URL(apiUrl);
-			HttpURLConnection con = (HttpURLConnection) url.openConnection();
-			con.setRequestMethod("POST");
-			con.setRequestProperty("Authorization", "Bearer " + authToken);
-			con.setRequestProperty("Content-Type", "application/x-www-form-urlencoded");
-			con.setRequestProperty("Content-Length", Integer.toString(postData.toString().getBytes().length));
-			con.setDoOutput(true);
-			con.getOutputStream().write(postData.toString().getBytes(StandardCharsets.UTF_8));
-			BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()));
-			String inputLine;
-			StringBuilder response = new StringBuilder();
-			while ((inputLine = in.readLine()) != null) {
-				response.append(inputLine);
-			}
-			in.close();
-
-			// Parse the JSON response to get the cost and time
-
-
-			JSONObject jsonObject = new JSONObject(response.toString());
-			JSONArray dataArray = jsonObject.getJSONArray("data");
-			JSONObject dataObject = dataArray.getJSONObject(0);
-			double cost = dataObject.getDouble("service_fee");
-
-
-			result = cost;
-			return  result;
-		} catch (Exception e) {
-			// TODO: handle exception
-		}
-		return result;
-
-	}
-
-	public static String getDateShip(int height, int lenght, int with,int weght,int  from_district_id, int from_ward_id,String tenHuyen, String tenXa) {
-		String apiUrl = "http://140.238.54.136/api/leadTime";
-		String authToken =token();
-		String result ="";
-		try {
-			int districtId_to = DistrictGHNUtils.getDistrictId(tenHuyen);
-			int ward_id_to = DistrictGHNUtils.getDistrictIdOfWard(tenXa,districtId_to);
-			Map<String, String> params = new HashMap<>();
-			params.put("height", height+"");
-			params.put("length", lenght+"");
-			params.put("width", with+"");
-			params.put("weight", weght+"");
-			params.put("from_district_id", from_district_id+"");
-			params.put("from_ward_id", from_ward_id+"");
-			params.put("to_district_id", districtId_to+"");
-			params.put("to_ward_id", ward_id_to+"");
-			StringBuilder postData = new StringBuilder();
-			for (Map.Entry<String, String> param : params.entrySet()) {
-				if (postData.length() != 0) postData.append('&');
-				postData.append(URLEncoder.encode(param.getKey(), "UTF-8"));
-				postData.append('=');
-				postData.append(URLEncoder.encode(param.getValue(), "UTF-8"));
-			}
-			URL url = new URL(apiUrl);
-			HttpURLConnection con = (HttpURLConnection) url.openConnection();
-			con.setRequestMethod("POST");
-			con.setRequestProperty("Authorization", "Bearer " + authToken);
-			con.setRequestProperty("Content-Type", "application/x-www-form-urlencoded");
-			con.setRequestProperty("Content-Length", Integer.toString(postData.toString().getBytes().length));
-			con.setDoOutput(true);
-			con.getOutputStream().write(postData.toString().getBytes(StandardCharsets.UTF_8));
-			BufferedReader in = new BufferedReader(new InputStreamReader(con.getInputStream()));
-			String inputLine;
-			StringBuilder response = new StringBuilder();
-			while ((inputLine = in.readLine()) != null) {
-				response.append(inputLine);
-			}
-			in.close();
-			System.out.println(response.toString());
-			JSONObject jsonObject = new JSONObject(response.toString());
-			JSONArray dataArray = jsonObject.getJSONArray("data");
-			JSONObject dataObject = dataArray.getJSONObject(0);
-
-			int time = dataObject.getInt("timestamp");
-			Date dateTime = new Date(time * 1000L); // Convert seconds to milliseconds
-			SimpleDateFormat dateFormat = new SimpleDateFormat("dd/MM/yyyy"); // Format the date
-			String formattedDate = dateFormat.format(dateTime); // Format the date as a string
-			Calendar calendar = Calendar.getInstance();
-			calendar.setTime(dateTime);
-			calendar.add(Calendar.DATE, 3); // Add 3 days
-			Date newdateTime = (Date) calendar.getTime();
-			SimpleDateFormat newDateFormat = new SimpleDateFormat("dd/MM/yyyy"); // Format the date
-			String newFormattedDate = newDateFormat.format(newdateTime); // Format the date as a string
-			result=(formattedDate+" - "+newFormattedDate);
-			return  result;
-
-
-
-		} catch (Exception e) {
-			// TODO: handle exception
-		}
-		return null;
-
-	}
 	public static String token()  {
 		try {
 			HttpClient httpClient = HttpClients.createDefault();
@@ -256,7 +133,7 @@ public class FeeGHNUtils {
 	}
 
 	public static void main(String[] args) {
-		System.out.println(getDateShip(10,10,13,200,2264,90816,"Quận Thủ Đức","Phường Linh Trung"));
+		System.out.println(feeAPIAndDate("100","100","100","1000",2264,90816,"Quận Thủ Đức","Phường Linh Trung").get(0));
 	}
 	
 }
